Refs #2898: Reduce code duplication in MDEvents.

+++ Code/Mantid/Framework/MDEvents/test/MDEventFactoryTest.h
@@ -59,7 +59,7 @@ public:
   }
 
 
-  int test_value;
+  size_t test_value;
 
 };
